Moved _renderItem to separate component
